fastapi/requirements.txt
annotated-doc==0.0.4
annotated-types==0.7.0
anyio==4.13.0
certifi==2026.2.25
click==8.3.2
colorama==0.4.6
fastapi==0.135.3
h11==0.16.0
idna==3.11
influxdb3-python==0.18.0
joblib==1.5.3
numpy==2.4.4
paho-mqtt==2.1.0
pandas==3.0.2
pyarrow==23.0.1
pydantic==2.12.5
pydantic_core==2.41.5
python-dateutil==2.9.0.post0
python-dotenv==1.2.2
reactivex==4.1.0
scikit-learn==1.7.2
scipy==1.17.1
six==1.17.0
starlette==1.0.0
threadpoolctl==3.6.0
typing-inspection==0.4.2
typing_extensions==4.15.0
tzdata==2026.1
urllib3==2.6.3
uvicorn==0.43.0
